+++ src/lib/functions/split_func.rs
@@ -12,6 +12,7 @@ use envmnt;
 pub(crate) fn invoke(function_args: &Vec<String>) -> Vec<String> {
     if function_args.len() != 2 {
         error!("split expects only 2 arguments (environment variable name, split by character)");
+        panic!("split expects only 2 arguments (environment variable name, split by character)");
     }
 
     let env_key = function_args[0].clone();
@@ -19,6 +20,7 @@ pub(crate) fn invoke(function_args: &Vec<String>) -> Vec<String> {
 
     if split_by.len() != 1 {
         error!("split expects a single character separator");
+        panic!("split expects a single character separator");
     }
 
     let split_by_char = split_by.chars().next().unwrap();